pyproject.toml
[tool.poetry]
name = "trquake"
version = "0.1.0"
description = ""
authors = ["ahmetkotan <ahmet.kotan@octoxlabs.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"

Django = {version = "^4.1.6"}
psycopg2-binary = "^2.9.5"
celery = "^5.2.7"
tweepy = "^4.12.1"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
redis = "^4.4.2"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
snscrape = {git = "https://github.com/JustAnotherArchivist/snscrape.git"}
pandas = "^1.5.3"
unidecode = "^1.3.6"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.1.0"
watchdog = {version = "^2.2.1", extras = ["watchmedo"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
